requirements.txt
flask==3.0.0
google-generativeai==0.8.3
sentence-transformers==3.0.1
chromadb==0.5.5
pandas==2.2.2
pillow==10.4.0
requests==2.32.3
numpy==1.26.4
werkzeug==3.0.4
